BOM_Observations/requirements.txt
amphoradata==0.9.2.dev1
requests
python-dotenv
pandas
numpy